renamed versiontabs into playground
